chore(localdebug): move save properties to env files into baseTunnelTaskTerminal
