+++ packages/vscode-extension/src/debug/taskTerminal/ngrokTunnelTaskTerminal.ts
@@ -12,12 +12,10 @@ import * as kill from "tree-kill";
 import * as util from "util";
 import * as vscode from "vscode";
 import { err, FxError, ok, Result, UserError, Void } from "@microsoft/teamsfx-api";
-import { envUtil, isV3Enabled } from "@microsoft/teamsfx-core";
+import { isV3Enabled } from "@microsoft/teamsfx-core";
 import { Correlator } from "@microsoft/teamsfx-core/build/common/correlator";
 import { DepsManager, DepsType } from "@microsoft/teamsfx-core/build/common/deps-checker";
 import { LocalEnvManager, TaskDefaultValue } from "@microsoft/teamsfx-core/build/common/local";
-import { pathUtils } from "@microsoft/teamsfx-core/build/component/utils/pathUtils";
-
 import VsCodeLogInstance from "../../commonlib/log";
 import { ExtensionErrors, ExtensionSource } from "../../error";
 import * as globalVariables from "../../globalVariables";
@@ -278,14 +276,6 @@ export class NgrokTunnelTaskTerminal extends BaseTunnelTaskTerminal {
 
   private async saveNgrokEndpointToEnv(endpoint: string): Promise<Result<OutputInfo, FxError>> {
     try {
-      const result: OutputInfo = {
-        file: undefined,
-        keys: [],
-      };
-      if (!isV3Enabled() || !globalVariables.workspaceUri?.fsPath || !this.args.env) {
-        return ok(result);
-      }
-
       const url = new URL(endpoint);
       const envVars: { [key: string]: string } = {};
       if (this.args?.output?.endpoint) {
@@ -294,25 +284,7 @@ export class NgrokTunnelTaskTerminal extends BaseTunnelTaskTerminal {
       if (this.args?.output?.domain) {
         envVars[this.args.output.domain] = url.hostname;
       }
-
-      if (Object.entries(envVars).length == 0) {
-        return ok(result);
-      }
-
-      result.keys = Object.keys(envVars);
-      const res = await envUtil.writeEnv(
-        globalVariables.workspaceUri.fsPath,
-        this.args.env,
-        envVars
-      );
-      const envFilePathResult = await pathUtils.getEnvFilePath(
-        globalVariables.workspaceUri.fsPath,
-        this.args.env
-      );
-      if (envFilePathResult.isOk()) {
-        result.file = envFilePathResult.value;
-      }
-      return res.isOk() ? ok(result) : err(res.error);
+      return this.savePropertiesToEnv(this.args.env, envVars);
     } catch (error: any) {
       return err(TunnelError.TunnelEnvError(error));
     }
